+++ custom_components/elektronny_gorod/__init__.py
@@ -30,6 +30,7 @@ PLATFORMS: list[Platform] = [
     Platform.CAMERA,
     Platform.LOCK,
     Platform.SENSOR,
+    Platform.SWITCH,
 ]